UnitTypes.jl

This package provides physical units as Julia types.

julia> using UnitTypes

julia> x = Meter(3)
3m

julia> typeof(x)
Meter

julia> typeof(x) <: AbstractLength
true

julia> typeof(x) <: AbstractCapacitance
false

This allows you to easily write functions with arguments restricted to variables having certain types.

julia> function goFaster(a::T) where T<:AbstractAcceleration end

This leads to correctness and very clear error messages.

julia> v = MeterPerSecond(3)
3m/s

julia> goFaster(v)
ERROR: MethodError: no method matching goFaster(::MeterPerSecond)

Closest candidates are:
  goFaster(::T) where T<:AbstractAcceleration

Introducing new types

Macros are used to introduce and create relationships around new types:

@makeBaseMeasure Torque NewtonMeter "N*m" - introduces a new basic Measure like Meter for Length or Meter3 Volume,

@deriveMeasure NewtonMeter(1) = MilliNewtonMeter(1000) "mN*m - introduces a new name for a Measure, often a prefix like Millimeter or an alternate name like Inch,

@makeDimension Diameter Meter - creates a Dimension, which is a Measure in some particular context, as diameter, radius, and circumference all refer to lengths of a circle.

Design

UnitTypes introduces an abstract type hierarchy of:

AbstractMeasure

Meter, Millimeter, ..., MeterPerSecond, MeterPerSecond2, ... See src/SIDerived.jl Inch, Foot, Mile, ..., See src/Imperial.jl

AbstractDimension - src/Dimension.jl

AbstractDiameter, AbstractRadius, ... AbstractDuration, ...

See src/typeTree.txt for a full list of the pre-defined types.

Please open an issue or PR to add more units.

As said, the idea is that a Measure is some quantity bearing units, while a Dimension is some context-specific application of a Measure. Within a Dimension multiple Measures may logically be used as long as they are dimensionally consistent. For instance, a circle may be described by its radius, diameter, or circumference, concepts that can be interchangeably converted, using any Measure of extent (<:AbstractLength). A function creating a circle can then internally store radii while accepting Radius, Diameter, or Circumference arguments, as the type system provides conversion between the argument and the function's internal convention.

Concrete Dimensions look like

struct Diameter{T <: AbstractLength } <: AbstractDimension
  value::T
end

and a concrete Measure is represented by

struct Meter <: AbstractLength
  value::Number
  toBase::Number
  unit::String
end

Measures.jl and the macros define the necessary convert()s and other operators. Please open an issue with a minimal working example if you run into conversion errors.

Logical operations

Using units correctly requires distinguishing between valid and invalid operations, which in some cases means not allowing convenient operations. Inches can be added, as can inch and millimeter, but only when computing area does inch*inch make sense. Inch * 3 is convenient while 3 / Inch is unlikely to be desirable.

With use and issues, these coherence rules will become more clear and explained by example.

Comparison with other packages

Unitful.jl

Unitful leverages parametric types to store units, giving flexibility at the cost of compile-time type uncertainty. It's two major limitations are the avoidance of angular measures, as they are not first-class but rather ratios, and rather lengthy type unions that clutter outputs, especially on error:

julia> function goSlower(x::T) where T<:Unitful.Acceleration end
goSlower (generic function with 1 method)

julia> a = 1u"mm"

julia> goSlower(a)
ERROR: MethodError: no method matching goSlower(::Quantity{Int64, 𝐋 , Unitful.FreeUnits{(mm,), 𝐋 , nothing}})

Closest candidates are:
  goSlower(::T) where T<:(Union{Quantity{T, 𝐋 𝐓^-2, U}, Level{L, S, Quantity{T, 𝐋 𝐓^-2, U}} where {L, S}} where {T, U}) 

As Unitful is the most widely used unit package, we provide a separate package ExchangeUnitful to enable interoperation with Unitful.

DynamicQuantities.jl

DynamicQuantities is newer and faster than Unitful because it "defines a simple statically-typed Quantity type for storing physical units." It does this by storing the exponents on the basic units, allowing any unit traceable to SI to be used. But this performant representation hurts readability, and while the unit representation may be able to be hidden behind overrides of show(), Julia is designed for types to be read and manipulated directly by users.

UnitTypes.jl

In the presence of Julia's type-heavy UI, these two, good attempts feel misdirected and motivate this package's literal typing of units. The limitation is that UnitTypes does not have a catch-all unit representation. Only units that have been defined by one of the macros may be represented, and complex units may need to have additional methods written to correctly convert between units, ie Celsius to Fahrenheit. See SIDerived.jl and Imperial.jl for examples.

Docs

UnitTypes.APerM — Type

This UnitType represents units of APerM.

UnitTypes.APerM2 — Type

This UnitType represents units of APerM2.

UnitTypes.Acre — Type

UnitType Acre is derived from Meter2, related by 4046.873, with supertype AbstractArea, and symbol [ac].

UnitTypes.Ampere — Type

This UnitType represents units of Ampere.

UnitTypes.Candela — Type

This UnitType represents units of Candela.

UnitTypes.CentiMeter — Type

UnitType CentiMeter is derived from Meter, related by 0.01, with supertype AbstractLength, and symbol [cm].

UnitTypes.Coulomb — Type

This UnitType represents units of Coulomb.

UnitTypes.Degree — Type

UnitType Degree is derived from Radian, related by 0.017453292519943295, with supertype AbstractAngle, and symbol [°].

UnitTypes.Farad — Type

This UnitType represents units of Farad.

UnitTypes.Femtometer — Type

UnitType Femtometer is derived from Meter, related by 1.0e-15, with supertype AbstractLength, and symbol [fm].

UnitTypes.FluidOunce — Type

UnitType FluidOunce is derived from Meter3, related by 0.0284130625, with supertype AbstractVolume, and symbol [floz].

UnitTypes.Foot — Type

UnitType Foot is derived from Meter, related by 0.30479999999999996, with supertype AbstractLength, and symbol [ft].

UnitTypes.FootPerSecond — Type

UnitType FootPerSecond is derived from MeterPerSecond, related by 0.30479999999999996, with supertype AbstractVelocity, and symbol [ft/s].

UnitTypes.Gallon — Type

UnitType Gallon is derived from Meter3, related by 4.54609, with supertype AbstractVolume, and symbol [gal].

UnitTypes.Henry — Type

This UnitType represents units of Henry.

UnitTypes.Hertz — Type

This UnitType represents units of Hertz.

UnitTypes.Inch — Type

UnitType Inch is derived from Meter, related by 0.0254, with supertype AbstractLength, and symbol [in].

UnitTypes.Kelvin — Type

This UnitType represents units of Kelvin.

UnitTypes.KgPerM2 — Type

This UnitType represents units of KgPerM2.

UnitTypes.KgPerM3 — Type

This UnitType represents units of KgPerM3.

UnitTypes.KiloGram — Type

This UnitType represents units of KiloGram.

UnitTypes.KiloMeter — Type

UnitType KiloMeter is derived from Meter, related by 1000.0, with supertype AbstractLength, and symbol [km].

UnitTypes.KiloNewton — Type

UnitType KiloNewton is derived from Newton, related by 1000.0, with supertype AbstractForce, and symbol [kN].

UnitTypes.KiloOhm — Type

UnitType KiloOhm is derived from Ohm, related by 1000.0, with supertype AbstractResistance, and symbol [kΩ].

UnitTypes.KiloVolt — Type

UnitType KiloVolt is derived from Volt, related by 1000.0, with supertype AbstractElectricPotential, and symbol [KV].

UnitTypes.Liter — Type

UnitType Liter is derived from Meter3, related by 0.001, with supertype AbstractVolume, and symbol [L].

UnitTypes.Lumen — Type

This UnitType represents units of Lumen.

UnitTypes.Lux — Type

This UnitType represents units of Lux.

UnitTypes.M3PerKg — Type

This UnitType represents units of M3PerKg.

UnitTypes.MegaOhm — Type

UnitType MegaOhm is derived from Ohm, related by 1.0e6, with supertype AbstractResistance, and symbol [MΩ].

UnitTypes.Meter — Type

This UnitType represents units of Meter.

UnitTypes.Meter2 — Type

This UnitType represents units of Meter2.

UnitTypes.Meter3 — Type

This UnitType represents units of Meter3.

UnitTypes.MeterPerSecond — Type

This UnitType represents units of MeterPerSecond.

UnitTypes.MeterPerSecond2 — Type

This UnitType represents units of MeterPerSecond2.

UnitTypes.MicroFarad — Type

UnitType MicroFarad is derived from Farad, related by 1.0e-6, with supertype AbstractCapacitance, and symbol [μF].

UnitTypes.MicroMeter — Type

UnitType MicroMeter is derived from Meter, related by 1.0e-6, with supertype AbstractLength, and symbol [μm].

UnitTypes.Mile — Type

UnitType Mile is derived from Meter, related by 1609.3439999999998, with supertype AbstractLength, and symbol [mi].

UnitTypes.MilliFarad — Type

UnitType MilliFarad is derived from Farad, related by 0.001, with supertype AbstractCapacitance, and symbol [mF].

UnitTypes.MilliHenry — Type

UnitType MilliHenry is derived from Henry, related by 0.001, with supertype AbstractInductance, and symbol [mH].

UnitTypes.MilliLiter — Type

UnitType MilliLiter is derived from Liter, related by 0.001, with supertype AbstractVolume, and symbol [mL].

UnitTypes.MilliMeter — Type

UnitType MilliMeter is derived from Meter, related by 0.001, with supertype AbstractLength, and symbol [mm].

UnitTypes.MilliNewton — Type

UnitType MilliNewton is derived from Newton, related by 0.001, with supertype AbstractForce, and symbol [mN].

UnitTypes.MilliNewtonMeter — Type

UnitType MilliNewtonMeter is derived from NewtonMeter, related by 0.001, with supertype AbstractTorque, and symbol [mN*m].

UnitTypes.MilliOhm — Type

UnitType MilliOhm is derived from Ohm, related by 0.001, with supertype AbstractResistance, and symbol [Ω].

UnitTypes.NanoFarad — Type

UnitType NanoFarad is derived from Farad, related by 1.0e-9, with supertype AbstractCapacitance, and symbol [nF].

UnitTypes.NanoMeter — Type

UnitType NanoMeter is derived from Meter, related by 1.0e-9, with supertype AbstractLength, and symbol [nm].

UnitTypes.NauticalMile — Type

UnitType NauticalMile is derived from Meter, related by 1852.0, with supertype AbstractLength, and symbol [nmi].

UnitTypes.Newton — Type

This UnitType represents units of Newton.

UnitTypes.NewtonMeter — Type

This UnitType represents units of NewtonMeter.

UnitTypes.NewtonMilliMeter — Type

UnitType NewtonMilliMeter is derived from NewtonMeter, related by 0.001, with supertype AbstractTorque, and symbol [N*mm].

UnitTypes.Ohm — Type

This UnitType represents units of Ohm.

UnitTypes.Ounce — Type

UnitType Ounce is derived from KiloGram, related by 0.028349523125, with supertype AbstractMass, and symbol [oz].

UnitTypes.Pascal — Type

This UnitType represents units of Pascal.

UnitTypes.PerSecond — Type

UnitType PerSecond is derived from Hertz, related by 1.0, with supertype AbstractFrequency, and symbol [s^-1].

UnitTypes.PicoFarad — Type

UnitType PicoFarad is derived from Farad, related by 1.0e-12, with supertype AbstractCapacitance, and symbol [pF].

UnitTypes.PicoMeter — Type

UnitType PicoMeter is derived from Meter, related by 1.0e-12, with supertype AbstractLength, and symbol [pm].

UnitTypes.Pint — Type

UnitType Pint is derived from Meter3, related by 0.56826126, with supertype AbstractVolume, and symbol [pt].

UnitTypes.PoundMass — Type

UnitType PoundMass is derived from KiloGram, related by 0.45359237, with supertype AbstractMass, and symbol [lbm].

UnitTypes.Quart — Type

UnitType Quart is derived from Meter3, related by 1.1365225, with supertype AbstractVolume, and symbol [qt].

UnitTypes.Radian — Type

This UnitType represents units of Radian.

UnitTypes.Second — Type

This UnitType represents units of Second.

UnitTypes.Siemens — Type

This UnitType represents units of Siemens.

UnitTypes.Slug — Type

UnitType Slug is derived from KiloGram, related by 14.59390294, with supertype AbstractMass, and symbol [slug].

UnitTypes.SquareFoot — Type

UnitType SquareFoot is derived from Meter2, related by 0.09290303999999998, with supertype AbstractArea, and symbol [sqft].

UnitTypes.SquareMile — Type

UnitType SquareMile is derived from Meter2, related by 2.5899881103359996e6, with supertype AbstractArea, and symbol [sqmi].

UnitTypes.Tesla — Type

This UnitType represents units of Tesla.

UnitTypes.Volt — Type

This UnitType represents units of Volt.

UnitTypes.Watt — Type

This UnitType represents units of Watt.

UnitTypes.Weber — Type

This UnitType represents units of Weber.

UnitTypes.Yard — Type

UnitType Yard is derived from Meter, related by 0.9143999999999999, with supertype AbstractLength, and symbol [yd].

Base.show — Method

@show functionality for Dimensions via dimension2String().

UnitTypes.dimension2String — Method

dimension2String(c::AbstractDimension) -> String

Returns a string representing dimenson c with format Module.DimensionName(value unit).

UnitTypes.@makeDimension — Macro

Make a new dimension dimName of measure; also creates 'AbstractdimName'

```julia @makeDimension Diameter Meter

d = Diameter(MilliMeter(3.4))
r = Radius(d)

```

UnitTypes.@relateDimensions — Macro

Defines various Base. functions that facilitate the given relationship. All types must already be defined and written in the form type1 = factor * type2, as in: julia @relateDimensions Diameter = 2.0*Radius